precompile: fail in (closer to) linear time
Do not allow serial compile fallbacks, since this causes performance and
output degredation when serial precompile fails. The precompilepkg
driver already ensured DAG ordering, so this work is not necessary or
useful to repeat. A bit of code cleanup and asking for Claude to
describe the argument list as well, plus a test to start to cover any of
this code functionality in CI.