inference: refactor data structures for inlining source caching
In inference, we now determine whether the source needed for inlining is
cached, and if not, we infer new source in local cache mode and
propagate it to inlining. The ultimate goal is to stop retrieving source
from the global code cache during inlining.

To achieve this, this commit comprehensively refactores the data
structures for passing source from inference to inlining. Specifically
it introduces the `InferredCallResult` type as a centralized data
structure for holding source information, which is stored in the
`info.call_results` field of `info::MethodMatchInfo`. The
`info.call_results` field is managed in the same way as `info.edges`,
making the code simpler.

It should be noted that `ConstCallInfo` has been removed. This would be
a breaking change for packages like Cthulhu, but since no information is
lost, preserving existing functionality should not be impossible.